Moved functions to base enough to get the base sample project parsed.
